+++ okdata/sdk/data/upload.py
@@ -1,4 +1,3 @@
-import requests
 import logging
 import os
 
@@ -14,14 +13,14 @@ class Upload(SDK):
         self.__name__ = "upload"
         super().__init__(config, auth)
 
-    def upload(self, fileName, datasetid, versionid, editionid):
+    def upload(self, fileName, datasetid, versionid, editionid, retries=0):
         url = self.config.get("s3BucketUrl")
         log.info(f"Uploading {fileName} to {datasetid} on: {url}")
         if url is None:
             raise KeyError("No s3 Bucket URL set")
 
         s3SignedData = self.create_s3_signed_data(
-            fileName, datasetid, versionid, editionid
+            fileName, datasetid, versionid, editionid, retries=retries
         )
         s3Data = {}
         if "message" in s3SignedData:
@@ -32,16 +31,19 @@ class Upload(SDK):
             s3Data[var] = s3SignedData["fields"][var]
 
         files = {"file": open(fileName, "rb")}
-        result = requests.post(url, data=s3Data, files=files)
+        upload_session = self.prepared_request_with_retries(retries=retries)
+        result = upload_session.post(url, data=s3Data, files=files)
         trace_id = s3SignedData.get("trace_id")
         data = {"result": result.status_code == 204, "trace_id": trace_id}
         return data
 
-    def create_s3_signed_data(self, fileName, datasetid, versionid, editionid):
+    def create_s3_signed_data(
+        self, fileName, datasetid, versionid, editionid, retries=0
+    ):
         edition = f"{datasetid}/{versionid}/{editionid}"
         data = {"filename": os.path.basename(fileName), "editionId": edition}
         url = self.config.get("uploadUrl")
         log.info(f"Creating s3 signed data with payload: {data} on: {url}")
         if url is None:
             raise KeyError("No Signed S3 URL set")
-        return self.post(url, data).json()
+        return self.post(url, data, retries=retries).json()